WIP: Three changes:
ItemControlBlock: Made ConstructDefaultItem optional. If given,
a explicit default to be created can be defined, else the
ConstructItem will be used to construct the default item.

ItemSet: Optimized to make the template methods as small as
possible due to tese being expanded by each type when compiled.
Also allowed me to move static fallback items completely
to *.cxx file.

SlotSet: Completely removed for now. The strategy is to solely
use ItemSet - under the vost that for each Item a type has to
be defined (by derivation). Something like SlotSet may also be
used in the future if a combination of the new Items and identifier
is needed for some reason. In principle, everyone is able to use
and handle ::BaseItem shared_ptr's as wanted - keep them in an
array or whetever. In principle there is no need for specialized
classes handling these - ::ItemSet is only for convenience and to
have a transition point

Change-Id: Ic6bfa9598a560acde22a124f6e3c893a595f9ac9